Fail on invalid argument
Prior to fixing c9s/CLIFramework#106, even if the command argument was invalid, it didn't cause the command to fail. After having this fixed, `InstallCommandTest::testUseCommand()` started failing. The `UseCommand` declared that it expected a non-prefixed with php- version name while it expected the build name instead.

The `testUseCommand()` is reworked to `testEnvCommand()` because this is what gets invoked from the shell wrapper when a user types `phpbrew use`.

Fixes #822.