use zsh chpwd hooks to look for .phpbrewrc
The original approach was looking for .phpbrewrc after EVERY command, when
in reality we only need to look for it when the current directory changes.
`trap` may be needed in bash, but zsh includes a hook mechanism for various
shell events, including changing directory.

This commit makes use of that hook mechanism to look for .phpbrewrc ONLY
when the current directory changes.

This affects only zsh users.

You can check http://zsh.sourceforge.net/Doc/Release/Functions.html#Hook-Functions
for more info on zsh hooks.