use php-vcr to record http interactions through all the test suite
This is an attempt to wrap every http interaction done while testing in a VCR cassette.
Not every single interaction can be wrapped since some of them download binary files
(which can't be stored in cassettes), but this commit wraps every interaction that
can be wrapped. This will allow for faster tests and less problems related to hitting
API limits, triggering anti-DDoS tools and such.

This also introduces two new phpunit group tags: cantVCR and macosIncompatible. The
first one is meant to be used in every test which by its own nature cannot be wrapped
into a VCR cassette. These are, for example, tests where binary files are downloaded.
The latter is just a treat to allow devs who use macOS to run the test suite without
tests breaking due to system libraries not available on that OS.

Now we can run something like vendor/bin/phpunit --exclude-group cantVCR whenever we
want to run the test suite without hitting the network. We can always do a full run
when needed.