Enforced PHP version in Composer configuration
This is needed in order to make sure we won't update locked dependencies to the versions that are not supported by the oldest supported PHP version while not using this verion during an update. Although, on CI we want to use the latest PHPUnit version that is supports the current PHP version.

The PHP 5.3.9 is used since this is the oldest PHP version that is supported by the currently locked dependencies (symfony/* specifically).