Removed the Machine and Process classes
The classes were used only for detection of the number of processors which was only used in a couple of tests while the actual code didn't use this number.

It is easier and more reliable to use MAKE_FLAGS that is controlled by the end user and is used by make consistenly across all tasks.