Made all patches part of the PHPBrew source tree
As of now, all patches should be included in the source tree to guarantee integrity. The checksum functionality is removed as no longer needed.

The patch sources are stored alongside the classes they are used by for better cohesion. Additionally, storing them as separate files would require creating a copy before applying anyways because the `patch` command wouldn't be able to use a file from within a PHAR. This issue is avoided by passing the patch to the CLI command via STDIN.