ci: Add a release workflow
Add a release workflow which builds the PHAR and, when triggered by a
release event, will sign and publish the PHAR.

:warning: This requires to:

- Replace `<TODO:privatekeyID>` by the actual GPG key ID (you can see it
  when running `$ gpg -K`) and `<TODO:email>` by the email address
associated with the GPG key used.
- You then need to add two environment variables `GPG_KEY_X` and
  `GPG_KEY_X_PASSPHRASE` to the GitHub Actions.

Note that the PHAR asc will be `phpbrew.phar.asc` instead of `phpbrew.asc`. I do not thing it is too important, but this follows the convention followed by [PHIVE](phar.io) which is widely adopted (at least by PHIVE itself, PHPUnit, Infection, PHP-Scoper & Box to name few).